+++ framework/CHANGELOG.md
@@ -140,6 +140,7 @@ Yii Framework 2 Change Log
 - Enh #10545: `yii\web\XMLResponseFormatter` changed to format models in a proper way (andrewnester)
 - Enh #10783: Added migration and unit-tests for `yii\i18n\DbMessageSource` (silverfire)
 - Enh #10797: Cleaned up requirements checker CSS (muhammadcahya)
+- Enh #10840: Added `yii\console\Controller::optionAliases()` method to support aliases for commands (pana1990)
 - Enh: Added last resort measure for `FileHelper::removeDirectory()` fail to unlink symlinks under Windows (samdark)
 - Enh: `AttributeBehavior::getValue()` now respects the callable in array format (silverfire)
 - Chg #6419: Added `yii\web\ErrorHandler::displayVars` make list of displayed vars customizable. `$_ENV` and `$_SERVER` are not displayed by default anymore (silverfire)